Split off decision_tree_model.hpp
